linux support started for block iterator

+++ src/include/storage/storage_defs.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #ifndef __APPLE__
+#include "numa.h"
 #include "numaif.h"
 #endif
 #include <algorithm>
@@ -183,7 +184,7 @@ class BlockAllocator {
       return new RawBlock();
     }
 #ifndef __APPLE__
-    return static_cast<RawBlock *>(numa_alloc_onnode(sizeof(RawBlock), region));
+    return static_cast<RawBlock *>(numa_alloc_onnode(sizeof(RawBlock), static_cast<int>(static_cast<int16_t>(region))));
 #else
     return new RawBlock();
 #endif
